requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
pandas==2.1.3
openpyxl==3.1.2

